+++ src/domain/journey/challenge/layout/ChallengePageBanner.tsx
@@ -1,12 +1,13 @@
-import React, { FC } from 'react';
+import React from 'react';
 import { useChallenge } from '../hooks/useChallenge';
 import { getVisualByType } from '../../../common/visual/utils/visuals.utils';
 import { VisualName } from '../../../common/visual/constants/visuals.constants';
 import useInnovationHubJourneyBannerRibbon from '../../../innovationHub/InnovationHubJourneyBannerRibbon/useInnovationHubJourneyBannerRibbon';
 import { useSpace } from '../../space/SpaceContext/useSpace';
 import ChildJourneyPageBanner from '../../common/childJourneyPageBanner/ChildJourneyPageBanner';
+import { BasePageBannerProps } from '../../common/EntityPageLayout/EntityPageLayoutTypes';
 
-const ChallengePageBanner: FC = () => {
+const ChallengePageBanner = (props: BasePageBannerProps) => {
   const { profile: spaceProfile } = useSpace();
   const { challenge, spaceId } = useChallenge();
   const banner = getVisualByType(VisualName.BANNER, spaceProfile?.visuals);
@@ -27,6 +28,7 @@ const ChallengePageBanner: FC = () => {
       tags={challenge?.profile.tagset?.tags}
       displayName={challenge?.profile.displayName ?? ''}
       tagline={challenge?.profile.tagline ?? ''}
+      {...props}
     />
   );
 };
