+++ src/domain/journey/opportunity/layout/OpportunityPageBanner.tsx
@@ -1,12 +1,13 @@
-import React, { FC } from 'react';
+import React from 'react';
 import { useOpportunity } from '../hooks/useOpportunity';
 import { getVisualByType } from '../../../common/visual/utils/visuals.utils';
 import { VisualName } from '../../../common/visual/constants/visuals.constants';
 import useInnovationHubJourneyBannerRibbon from '../../../innovationHub/InnovationHubJourneyBannerRibbon/useInnovationHubJourneyBannerRibbon';
 import ChildJourneyPageBanner from '../../common/childJourneyPageBanner/ChildJourneyPageBanner';
 import { useSpace } from '../../space/SpaceContext/useSpace';
+import { BasePageBannerProps } from '../../common/EntityPageLayout/EntityPageLayoutTypes';
 
-const OpportunityPageBanner: FC = () => {
+const OpportunityPageBanner = (props: BasePageBannerProps) => {
   const { profile: spaceProfile } = useSpace();
   const { opportunity, spaceId } = useOpportunity();
   const banner = getVisualByType(VisualName.BANNER, spaceProfile?.visuals);
@@ -27,6 +28,7 @@ const OpportunityPageBanner: FC = () => {
       tags={opportunity?.profile.tagset?.tags}
       displayName={opportunity?.profile.displayName ?? ''}
       tagline={opportunity?.profile.tagline ?? ''}
+      {...props}
     />
   );
 };
