+++ src/domain/journey/space/SpacePageTabs.tsx
@@ -31,6 +31,7 @@ import { useNavigate } from 'react-router-dom';
 import getEntityColor from '../../shared/utils/getEntityColor';
 import useShare from '../../../core/utils/Share';
 import { EntityTabsProps } from '../common/EntityPageLayout';
+import { gutters } from '../../../core/ui/grid/utils';
 
 interface TabDefinition {
   label: ReactNode;
@@ -97,7 +98,7 @@ const SpacePageTabs: FC<EntityPageTabsProps> = ({
   if (mobile) {
     return (
       <>
-        <Paper sx={{ position: 'fixed', bottom: 0, left: 0, right: 0 }} elevation={3} square>
+        <Paper sx={{ position: 'fixed', bottom: 0, left: 0, right: 0, paddingBottom: gutters() }} elevation={3} square>
           <BottomNavigation
             showLabels
             value={currentTab}
