+++ src/main/ui/layout/TopLevelDesktopLayout.tsx
@@ -3,17 +3,21 @@ import PlatformNavigationBar, { PlatformNavigationBarProps } from '../platformNa
 import Footer from '../platformFooter/PlatformFooter';
 import FloatingActionButtons from '../../../core/ui/button/FloatingActionButtons';
 import PlatformHelpButton from '../helpButton/PlatformHelpButton';
+import PoweredBy from '../poweredBy/PoweredBy';
+import { Box, Paper } from '@mui/material';
 
 interface TopLevelDesktopLayoutProps {
   breadcrumbs?: PlatformNavigationBarProps['breadcrumbs'];
   header?: ReactNode;
   floatingActions?: ReactNode; // Defaults to the HelpButton
+  addWatermark?: boolean;
 }
 
 const TopLevelDesktopLayout = ({
   breadcrumbs,
   header,
   floatingActions,
+  addWatermark,
   children,
 }: PropsWithChildren<TopLevelDesktopLayoutProps>) => {
   return (
@@ -23,6 +27,11 @@ const TopLevelDesktopLayout = ({
       {children}
       <Footer />
       {floatingActions ?? <FloatingActionButtons floatingActions={<PlatformHelpButton />} />}
+      {addWatermark && (
+        <Box component={Paper} square position="fixed" bottom={0} left={0} right={0}>
+          <PoweredBy compact />
+        </Box>
+      )}
     </>
   );
 };
